PythanctuaryRPG/requirements.txt
numpy
opencv-python
Cython
IPython
pyautogui
Pillow
pytesseract